adapted Add Item component to be reusable for other purposes with callback and submit label from props
